move calculateSNR to util.py
